feat: LecuterHandler.java 강의열람 및 강의조회, 강의수강신청, 컨텐츠 조회, 컨텐츠 열람 구현 schema.sql : Enrolment 로 수정

+++ src/main/java/com/example/onlineeduplatformlecture/service/impl/LectureServiceImpl.java
@@ -25,7 +25,7 @@ public class LectureServiceImpl implements LectureService {
     // GET /lectures/{lectureId} (강의 열람)
 
     @Override
-    public Mono<Lecture> getLecture(int lectureId) {
+    public Mono<Lecture> getLecture(Long lectureId) {
         return lectureRepository.findByLectureId(lectureId);
     }
